[Feature] Add CommandProcessor Draft
