+++ GreenLight/ControlTypes.h
@@ -1,0 +1,29 @@
+#pragma once
+#include "IPrinter.h"
+
+enum class CommandType {
+	ADD,
+	DEL,
+	SCH,
+	MOD,
+	INVALID
+};
+
+enum class FilterOption {
+	FirstName,
+	LastName,
+	MiddlePhoneNum,
+	LastPhoneNum,
+	BirthdayYear,
+	BirthdayMonth,
+	BirthdayDay,
+	None,
+	Invalid
+};
+
+struct Command {
+	CommandType type_;
+	PrintOption printOption_;
+	FilterOption filterOption_;
+	std::vector<std::string> arguments_;
+};
