+++ GreenLight/GreenLight.vcxproj
@@ -140,11 +140,14 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="CommandParser.cpp" />
+    <ClCompile Include="CommandProcessor.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="MemoryDatabase.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="CommandParser.h" />
+    <ClInclude Include="CommandProcessor.h" />
+    <ClInclude Include="ControlTypes.h" />
     <ClInclude Include="IDatabase.h" />
     <ClInclude Include="IPrinter.h" />
     <ClInclude Include="IOutput.h" />
