+++ Test/Test.vcxproj
@@ -37,6 +37,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="CommandParserTest.cpp" />
+    <ClCompile Include="CommandProcessorTest.cpp" />
     <ClCompile Include="MemoryDatabaseTest.cpp" />
     <ClCompile Include="test.cpp" />
     <ClCompile Include="pch.cpp">
