+++ Test/Test.vcxproj.filters
@@ -9,6 +9,9 @@
     <ClCompile Include="MemoryDatabaseTest.cpp">
       <Filter>Model</Filter>
     </ClCompile>
+    <ClCompile Include="CommandProcessorTest.cpp">
+      <Filter>Control</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
